requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
jose[cryptography]
python-jose
pydantic[email]
python-dotenv
aiosmtplib
jinja2